Rename files and add info in storie
